Move to https://api.crypto2099.io/v1/sigma
this move involves using the new Crypto2099 endpoint in order to fetch the epoch nonce.
While the new api now also allows for you to get teh sigma based on your pool ID this container
will continue to calculate the sigma on its own for the time being. Reason for that is mainly
that it was less to change and on another side I am debating keeping maintaining those scripts
for at least a shoer while.